wrlinux-common.inc: don't override hostname value
Issue: LIN9-2283

We are now seeing weird login hostname on qemu machines. Console output
are like below.

  Wind River Linux 9.0.0.0 192.168.7.2 ttyS0

  192 login:

This is because we override the 'hostname' value in wrlinux-common.inc
file, resulting in no default /etc/hostname created.

As far as I can see, ${MACHINE} is a reasonable default value for hostname
than just leaving it empty.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>