features: dpdk: update to make suitable for LIN10
Issue: LIN10-708

Since we no longer have wr-dpdk and DPDK versions have been updated we
have to make a few updates to ensure the documentation and
configuration matches what is being delivered.

Since the DPDK recipes are found in meta-dpdk we also take this
opportunity to add 'dpdk' as a layerrecommend.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>