feature/target-toolchain: Update README file and adjust index values
The README file was updated to include more information about what is
actually included when this template is enabled.  The template has been
supported on non-x86 machines for a while now.  So clarify this by removing
the support statement.

Finally add a keyword identifier to the file so that the layer index
can get a proper description of the feature.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>