wrlinux[-small]: use default oe-core DISTRO_FEATURES [ TA195552 ]
Issue: TA195552

Add the 3g, acl, nfc, zeroconf DISTRO_FEATURES for the wrlinux and
wrlinux-small distros.

In WRLinux-5, the 3g, acl, nfc DISTRO_FEATURES (DFs) were not used
since typical deployments did not require them. 3g and nfc are
now more common and acl is a requirement for kvm and other
sub-systems so it makes sense to use the default oe-core DFs.
Graphics-related distro features are added for the sato distro.

buildhistory-diff shows that for a wrlinux-small, world build,
the following package's configuration and therefore size
changed:
   autofs
   connman
   coreutils
   groff-doc
   libarchive
   logrotate
   pulseaudio-lib-bluez5-util
   samhain-client
   samhain-standalone
   tcl
   vim-tiny
   vim-common
The size of most of these pacakges grew by less than 1%.
Most packages were adding acl support but connman added ofono
as a dependency due to the 3g DF as expected.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>