wrlinux-whitelist.conf: remove workspacelayer from PNWHITELIST_LAYERS
Issue: LIN10-1947

WRLinux has a mechanism to strict a layer expose to users only a specified
list of recipes. This is done via PNWHITELIST_LAYERS and PNWHITELIST.

The devtool tool would create a layer called workspacelayer itself to
manage its user recipes. In order for devtool to work correctly in WRLinux,
we need to remove 'workspacelayer' from PNWHITELIST_LAYERS.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>