feature/edac: drop kernel fragment and issue warning
Issue: LIN10-2327

An error in kernel EDAC patches results in
patch application failure when feature/edac is added:
    ERROR]: Application of .kernel-meta//patches//features/edac/EDAC-pnd2-Return-proper-error-value-from-apl_rd_reg.patch failed.
Omit the scc fragment and issue a warning to the user.

The edac-utils package is required by a our images so adding
feature/edac will still add it even though it will not work
properly.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>