wrlinux-common.inc: delete the WRSDK_RDEPENDS variable
Issue: LIN1018-610

This variable is too hacky and is now causing problem. So I choose
to delete it.

From the name of the variable, I can see that it wants to add
some nativesdk tools to the traditional SDK. This could be better
achieved by modifying nativesdk-packagegroup-sdk-host.bb file
or adding a corresponding bbappend file.

Maybe it wants to also change buildtools-tarball, but it should
also be done via bbappend file, or just modify the bb file if
upstream accepts it.

Anyway, if the nativesdk tool is not a generic requirement, the
user should add it via TOOLCHAIN_HOST_TASK.

From the list of what WRSDK_RDEPENDS has, the only two items that
might make some sense are "nativesdk-mtools" and "nativesdk-syslinux".

These two items are added via bbappend file of nativesdk-packagegroup-sdk-host.

For this eSDK installation bug, another way to fix it is to add nativesdk
perl modules to WRSDK_RDEPENDS, e.g. nativesdk-perl-module-data-dumper.
But I don't think we shoud continue this hacky solution.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>