lmbench: add warning when no enough space for lmdd
Issue: LIN10-4267

The lmbench-test will create a file in ${FSDIR} using lmdd. If there is no
enough space in that directory for creating the file, lmdd only outputs a
simple message like below:

	write: wanted=8192 got=4096

The message is too implicit for debugging. So we add additional checking
of file size after creating the file, if the file size is smaller then
expected, we output a warning message to inform user of noticing the free
space of the directory.

(LOCAL REV: NOT UPSTREAM) -- Pending.

Signed-off-by: Rui Wang <rui.wang@windriver.com>

Issue: LIN1018-1772

Signed-off-by: Dengke Du <dengke.du@windriver.com>