wrlinux-uninative.inc: move tarballs into a directory named by checksum
Issue: LIN1018-1802

The local fetcher (file://) doesn't verify checksum, which means that there
is no error when yocto-uninative.inc is updated, this is not what we need.
Put checksum into UNINATIVE_URL, and move the tarball into a directory named
by checksum, so that we will get errors when it is updated.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>