feature/kernel-alt-headers: Allow kernel specific userspace headers
Issue: LIN10-2604

In the prior version of Wind River Linux, there was a way to enable
userspace headers to be produced to match the kernel sources.  While this
is not recommended, it is something the users may want to do.  See:
  oe-core/meta/recipes-kernel/linux-libc-headers/linux-libc-headers.inc

The new template also automates this process and will trigger installation
of these headers into /usr-alt/include.  (Note the package is not installed
into a target rootfs by default, only into an SDK.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>

Issue: LIN1018-140

Adjust file names for linux-yocto-dev and linux-yocto_4.18

Signed-off-by: He Zhe <zhe.he@windriver.com>