valgrind: move patch dir to match new version
Issue: LIN1019-8

Valgrind was upreved in oe-core and we rebased last night.
We have two valgrind patches in wrlinux applied by a version
specific bbappend so the build was broken when the rebase was pushed.
I already fixed the bbappend version in:
   f0b0cdf valgrind: rename bbappend to match upstream
but I didn't do a test build locally so the patch stage is
still failing. This change fixes that problem.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>