conf/wrlinux-blacklist.conf: Make the kernel blacklist dynamic
We want to make sure that an unexpected kernel is not built.  This includes an
alternative kernel that claims to support 'virtual/kernel', or a version that
we do not expect.

Both checks will verify that a PREFERRED_PROVIDER or PREFERRED_VERSION is
defined prior to blacklisting items.  This makes the system the most flexible
by allowing a user to let the system choose if necessary.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>