Deprecate qemuarma9, link this to qemuarm.
Issue: LIN1019-50

qemuarma9 is no longer supported, this is simply a compatibility shim to
the a15 version.  These are compatible from a userspace perspective.

The historical qemuarma9 is based on the arm versatile hardware emulation,
but the point of our QEMU architectures is not platform emulation but
ability for overall system simulation, and with a focus on userspace
testing.  The qemuarm uses the generic '-machine virt' instead, which is
more efficient than the versatile BSP.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>