vrf: enable systemd-tmpfiles.service for unionfs
Issue: LIN1019-1802

This issue is same with LIN1019-1657. Since systemd 240, it will check
pid file of service in /run even if the PIDFile of service points to
/var/run. So the separated directories /run and /var/run will make
systemd can not check the pid file and get the correct status of
services.

We don't need to create /var/run in rootfs layout since it should be
symbolic link of /run. So enable systemd-tmpfiles-setup.service to
create it when booting for unionfs.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>