wrlinux distro configuration: share the sample key variables
The feature/ostree template needs access to the same variables that
are in the security templates in order to implement key signing.  The
key directory variables have been moved to a common location so they
can be used by any template.

The RPM_GPG_NAME should have always been a ?= and not an = so it can
be overridden by the end user.

[ Issue: LIN1018-4565 ]
[ Issue: LIN1018-4601 ]
[ Issue: LIN1018-4566 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>