feature/ostree: Fix variable usage
MODULE_TARBALL_DEPLOY, IMAGE_BOOT_FILES_x86, and
IMAGE_BOOT_FILES_x86-64 should not be changed by the feature/ostree.
These are left over from the prior integration of ostree and do not
impact the function of the ostree integration.

The WKS_FILE needs to be force set by the template in order for wic to
work properly.  The OSTREE_WKS_FILE can be set to a custom .wks or
.wks.in file if desired in the local.conf.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>