OV: dpdk: Remove preference to version 17
Issue: LIN1019-2673

openvswith used not to be able to work with dpdk v18 and thus we specified the
preferred version as v17 for OV.

Since openvswitch v2.11.0, it has been able to work with DPDK 18.11 and later
versions. See https://www.openvswitch.org/releases/NEWS-2.11.0.txt

As we have changed to openvswitch v2.11. It's time to remove the preference.

Signed-off-by: He Zhe <zhe.he@windriver.com>