wrlinux-overc: depend on wr-ostree
Issue: LIN1019-2676

In wrlinux-x, the commit 51155762bc23 [data/xml/wr-ostree-dl: Add
ostree support], moved the ostree downloads from meta-overc-dl to
ostree-dl as these are not only used by overc but also by other
configurations, thereby eliminating duplicates.

We could just have meta-overc-dl depend on ostree-dl as the above
commit indicates but although this would share the download it could
result in multiple ostree configurations or even run into version
issues. So make wr-ostree a LAYERDEPENDS, allowing not only the
download but also the configuration to be shared. If for some reason
it is not desirable to share the configuration, the bblayers.conf will
have to be manipulated manually to remove the wr-ostree layer from the
list.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>