wrlinux-overc: be the first to 'soft' assign virtual/kernel
Issue: LINUXEXEC-2883

The upstream meta-overc layer sets a 'soft' assignment of
virtual/kernel to linux-yocto-dev, we however don't ever want this to
be the case. We therefor include our own 'soft' assignment of
virtual/kernel to 'linux-yocto'. This still leave template features
and other configurations to use a hard assignment which they would
want to use anyways to overcome the 'soft' assignment in meta-overc.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>