wrlinux-overc: get qemuarm64 running properly
Issue: LIN1019-2964

While testing the work on OverC for qemuarm64 we noticed several
issues. The VM would not always run smoothly due to tight memory
constraints, here we update the README to specify 2GB of memory to the
VM.

Since the cube-server has device.mgr==self the hooks does not setup
the resolv.conf and we need to re-enable the systemd-resolved.service
to do this work. We therefor reinstate this service, only for
cube-server, as part of the installer configuration.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>