overc: immutable.py: add support for SIGGEN_UNLOCKED_RECIPES
Issue: LIN1019-3325

Following the official docs to run the overc-install to build an overc
image or the documentation in README.qemuarm64 with signature locking
active we get an error when running 'bitbake build-sysroots':

    ERROR: The build-sysroots:do_build_target_sysroot task does not have a defined signature.

Unfortunately 'build-sysroots' is treated differently than images and
as such we can't simply add it to the list of BITBAKE_TARGETS in the
release scripts used to generate the pr-lockdown files. Fortunately we
actually don't care about the signature for this recipe so we can
simply 'unlock' the recipe. Here we replicate the 'unlocked' recipe
handlers from SignatureGeneratorOEBasicHashMixIn in Poky, and likewise
make use of the SIGGEN_UNLOCKED_RECIPES variable to allow the list of
unlocked recipes to be passed.

Setting a default value for SIGGEN_UNLOCKED_RECIPES will  prevent
folks from always having to add this to their local.conf for OverC
builds.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>