common native package list: update the generation script.
Issue: LINCD-438

Update the gen_common_native_package_list.sh script.
The format for:
$ ./wrlinux-x/setup.sh --list-machines was changed by:
   d92eeab44f completion output of 'setup.sh --list-distros/--list-templates/--list-machines'
so fix the 'list_machines' funciton to deal with that.

Update wr-common-packages-native.inc based on the updated script.
This is a clone of the wr-common-packages-native.inc from LIN1019,
mereged to master-wr to test if the automatic update mechanism works.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>