linux-yocto-dev: qemumips64: Revert patches using generic VDSO
Issue: LINCD-602

This series of patches turn to use generic VDSO implementation which does not
work and causes boot failure for mips64 architectures.

Yocto kernel maintainer is working on this issue. But we cannot wait as we're
getting close to CD release. This should be dropped once it's fixed upstream.

Signed-off-by: He Zhe <zhe.he@windriver.com>