qemu:Remove "_append_" for the PACKAGECONFIG of qemu
Issue: LINCD-1758

The spice and rbd are appended to the variable PACKAGECONFIG with the following
configuration in qemu-ovp.inc.

"""
PACKAGECONFIG_append_x86-64_class-target = " spice rbd"
"""
If we would like to disable the ceph(rbd) or remove it, we are obliged to
get rid of " rbd" in local configuration.

""""
PACKAGECONFIG_remove_x86-64_class-target = " rbd"

""""

In this case, it would casue a hack, So we need to make the changes in order to
affect a long term solution

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
Signed-off-by: He Zhe <zhe.he@windriver.com>