bcm-2xxx-rpi: Add temporary kernel patches for binary distrobution
Issue: LINUXEXEC-8353

This will enable us to boot the RPi3 and the RPi4 as well as see the HDMI
console and serial console at the same time while installing and using the
device. It is a major improvement over what we had before, because you can
interact with the u-boot on the HDMI console using a USB keyboard.

Signed-off-by: He Zhe <zhe.he@windriver.com>