wrlinux: lxdm-theme: upgrade the resolution of wallpapers to 4k
Issue: LINCD-3093

The resolution of orininal wallpapers pictures is 1280*840, when connect
to 4k display monitor, the screen is splited, there are 9 little pictures
displayed on 4k monitor. So, create 4k resolution wallpapers to
fix this issue.

Signed-off-by: Meng Li <Meng.Li@windriver.com>

The previous commits was overwritten by
c82e616545c929d1738927cac2648a6ee8086655, so merge again to fix it.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>