feature/passwd-expire: Add template
Issue: LINUXEXEC-10554

Add template feature/passwd-expire to provide a mechanism to force users
to update their passwords when next time serial login.

Introduce 2 variables. One is FORCE_PASSWD_EXPIRE which is the flag to
enable this feature, and it should not be disabled. The other one is
FORCE_EXPIRED_USERS which works with EXTRA_USERS_PARAMS from
extrausers.bbclass. Users set in FORCE_EXPIRED_USERS will be prompted to
update the passwords when next time serial login.

Signed-off-by: Kai Kang <kai.kang@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>