CUT: specify inode size to 128 byte
Issue: LINCD-6593

After e2fsprogs upgrades to 1.46.4, the defaults for mke2fs
 now call for 256 byte inodes for all file systems (with
the exception of file systems for the GNU Hurd, which only
supports 128 byte inodes) [1].

So when use lvcreate to claim 50M logical volume as below:
 # lvcreate -L 50 -n lv_test1 vg_test && mke2fs /dev/vg_test/lv_test1

And continue to check the file system size of the logical volume.
 # df -h | grep dev/mapper/vg_test-lv_test1
 /dev/mapper/vg_test-lv_test1   48M   14K   46M   1% /mnt/lv-test

So set the inode size to 128 bytes to fix the gap.

[1] https://git.kernel.org/pub/scm/fs/ext2/e2fsprogs.git/tree/doc/RelNotes/v1.46.4.txt

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>