wrlinux-kernel-dev: don't hotpatch for SDK Makefile target
Issue: LINCD-6583

In commit 9b77e2d65c4e6842d74495337119d5bcfb18dae7
("Revert "linux-yocto-dev: support to build under SDK")

we dropped the hotpatch to the toplevel kernel Makefile since we
simply can't be doing that for multiple reasons.

As it turns out, the patch itself was shared between linux-yocto
and linux-yocto-dev (but not via the same commit) so that is yet
another reason why it can't and shouldn't be done here.

So drop the 2nd reference as well.  Add notices to deter anyone from
trying to hotpatch the kernel from here ever again.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>