qemuarm64: Only apply a patch for kernel v5.10
Issue: LINCD-6882

This patch is only for v5.10 kernel, but when the project is setup
with nodistro, the qemuarm64 kernel will translate to v5.14. You know
configures in linux-yocto-extra-kernel-src.inc are provided only when
"OVERRIDES_append = ":osv-wrlinux"" is set, and nodistro should not
have this. However nodistro set it in local.conf currently, may user
space will fix it in the future. Now we add an if statement to avoid
apply this patch for non 5.10 kernel as workaround.

Signed-off-by: Yanfei Xu <yanfei.xu@windriver.com>