linux-libc-headers-custom: Prepend KERNEL_HEADER_DIR to includedir
Issue: LINCD-8088

To avoid conflict with linux-libc-headers, linux-libc-headers-custom usually
puts kernel headers in /usr-alt rather then /usr. And thus oe_multilib_header
would fail as it finds headers in $includedir which is /usr by default.

This patch replaces /usr with /usr-alt in and only in this recipe.

Add default assignment to KERNEL_HEADER_DIR so that linux-libc-headers-custom
can be built by default.

Signed-off-by: He Zhe <zhe.he@windriver.com>